move functions to traits to allow any version of phpunit
