+++ src/Traits/HasUtilityMethodsTrait.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace Omnipay\Tests\Traits;
+
+trait HasUtilityMethodsTrait
+{
+
+
+    /**
+     * Converts a string to camel case
+     *
+     * @param string $str
+     * @return string
+     */
+    public function camelCase($str)
+    {
+        return preg_replace_callback(
+            '/_([a-z])/',
+            function ($match) {
+                return strtoupper($match[1]);
+            },
+            $str
+        );
+    }
+
+    /**
+     * Helper method used by gateway test classes to generate a valid test credit card
+     */
+    public function getValidCard()
+    {
+        return [
+            'firstName' => 'Example',
+            'lastName' => 'User',
+            'number' => '4111111111111111',
+            'expiryMonth' => rand(1, 12),
+            'expiryYear' => gmdate('Y') + rand(1, 5),
+            'cvv' => rand(100, 999),
+            'billingAddress1' => '123 Billing St',
+            'billingAddress2' => 'Billsville',
+            'billingCity' => 'Billstown',
+            'billingPostcode' => '12345',
+            'billingState' => 'CA',
+            'billingCountry' => 'US',
+            'billingPhone' => '(555) 123-4567',
+            'shippingAddress1' => '123 Shipping St',
+            'shippingAddress2' => 'Shipsville',
+            'shippingCity' => 'Shipstown',
+            'shippingPostcode' => '54321',
+            'shippingState' => 'NY',
+            'shippingCountry' => 'US',
+            'shippingPhone' => '(555) 987-6543',
+        ];
+    }
+}
